Revisit parsing functions and their const correctness
I had not checked before how the parsing API in common.h worked for
const or non const variables. In closer inspection, the API causes more
problems than those it solves. Better stop using these functions as soon
as possible, and not extend it to the point where it will be harder to
change.