Start using libxkbcommon to interpret .xkb files
Even though in the future we will need to store our own state for a
loaded keymap (as libxkbcommon does not provide an API for modifying a
keymap). We always want the rendering of the keyboard to be done through
it to ensure that what we see is what libxkbcommon will interpret
everywhere in the system.

Also added a function to reverse what xkb_keymap_install() does. This
way we can work from already installed layouts without knowing the
source file from which it was installed.