Use keysyms computed by libxkbcommon as key labels
For the default keyboard rendering we want to show what the user would
get when pressing that key. For this we translate the keycode using
libxkbcommon and show it.

This translation using libxkbcommon is expected to have lots of special
cases. Sometimes we want the utf8 representation (english letters),
others we want the name returned by libxkbcommon (modifier keys) and
sometimes we want our own representation (Fn key). And I haven't checked
what happens for combining characters from Unicode.