Add a test .xkb file with few components
As far as I can tell most of the documentation for the syntax of .xkb
files has been lost in time. This is why we may have to learn this by a
combination of trial and error, and reading libxkbcommon's source code.

Here I'm starting a small test layout that has less clutter that what we
were using before, so the meaning of the syntax is more clear.