Put together the state of keyboard_view into a FSM
Managing all the state of the keyboard_view across multiple signal
handlers was becoming nasty. Having to jump between handlers to
understand the workflow takes a lot of time.

Instead, now all handlers call the same function kv_update() whose task
is to uptate a finite state machine for the keyboard_view. I think this
will ease the readability and debuggability of the code.