Modify behavior of repeated keycode editing
The way we handled assigning a keycode that is already being used was to
cancel the process. Currently, the most useful use case is to allow the
user to do a sequence of click, key press, click, key press, etc. for
keycode remapping. The current handling of double assignment of keycodes
does not work well with this use case.

Instead of this, we now allow the assignment to happen, and instead mark
the key that had the same keycode as unassigned. Unassigned keys will be
rendered without label to provide feedback to the user.