Snap split to increments of 0.125 (1/8)
From the user's perspective this keeps the rendering pixel perfect,
allows the user to perfectly align things and makes it easy to have
squared keyboards.

Because these increments are exactly representable by a floating point
number we can safely test for equality. This prepares the code to detect
alignments and make more complex operations when needed. For instance,
when resizing a segment of a multirow key if it now becomes a
rectangular key, we need to change the type of the segment to now
inherith the width.