Implement resizing of the right edge of rectangular keys
This is the simplest form of resizing where we just change the key's
width and recompute the glue normally.

Turns out the pinning API that was thought of in the previous commit
isn't really necessary. when the keyboard changes size thigs don't go
crazy. The keyboard is recentered which makes it offset a bit from the
clicked position, but it feels good enough.

Also while implementing this 2 bugs were fixed, first the keyboard size
computation didn't take into account glue and multirow key widths.
Second, we were not correctly computing the width of rectangular
multirow keys from their parent.