Implement partial left edge resizing and snapshot creation
I made some progress with the resize key implementation:

Compute the delta between the first click and the current pointer
position, instead of making an absolute distance calculation with the
key rectangle.

Implement user glue to have a way of specifying glue that won't be
altered by the glue computation. At the moment internal glue and user
glue are not mutually exclusive, but having both causes issues. This
leads me to think mabe we should enforce this.

Implemented a way to make a partial copy of the keyboard database, called
size snapshots. It clones the rows linked list and each key linked list
but without cloning the multirow cyclic linked lists.

Stub implementation of kv_right_align_changed_glue(), this function
should compare an old glue state and a newer one, and be able to move
the postion of the new glue to emulate a right "gravity". This is the
opposite of what happens automatically when resizing a right edge of a
key and recomputing the glue.