Finish implementation of the left edge key resize
The previous idea of making internal glue and user glue be mutually
exclusive has been dropped. Instead the semantics are that user glue
will always be located in multirow parents, and it may be combined with
internal glue, in which case the effective glue is the addition of both
glues. This user glue is then included in the internal glue computation
for the multirow segments of that multirow key. This means that changing
the user glue in a multirow key must trigger a recomputation of the
internal glue. The reason for not allowing user glue in other multirow
segments is that the semantics would get weird if multiple segments have
user glue, and now that we agree it should be unique the best location
is the multirow parent.

After a lot of thinking I have decided to drop the snapshot approach.
The original idea was to compute how the glue changed after resizing the
key and then reposition this glue to simulate keys being attracted to
the right. Even though computing how much glue changed is easy,
computing _where_ it should be added is not so straight forward because
depending on the ammount of glue it may need to be propagated to
different places in the row, or maybe even other row. I'm guessing this
algorithm will look like the glue computation algorithm but that instead
looks for dependencues between the keys starting from the one being
resized.

Another approach that I thought about was transforming the keyboard data
structure to a mirrored version, where the row lists start at the
rightmost key. The problem now is to decide which internal glue should
become user glue in it's mirrored version, ensuring the mirrored version
renders exactly the same. These glue transformations seem quite
involved, even appearing to be equivalent to the dependency search
above. Also, we don't have a right aligned rendering function so we
would have to debug things manually or implement this rendering function.

Because I could not come up with a simple enough solution, and this does
not seem to be worth the development time, I decided to fake it. What we
do is always modify the user glue of the key being resized. This has
the drawback of creating a gap before the resized key when it's
shrunken, and keys that could move right will not do so. I'm fine with
this as there will be a move tool to fill these gaps. The only ugly
quirk happens when resizing the left edge of a key that has internal
glue to the right (in the next key), in this case the left edge won't
move as resizing the key will first consume the internal glue. As an
example in the current test layout, try resizing the left edge of the
key with keycode 2.

There may be ways of fixing this. I thought about making the consumed
internal glue, user glue but as the semantics for user glue have settled
to be added on top of internal glue, it won't work (there will always be
internal glue). An actual solution may involve adding a "glue fixing"
mechanism or a third type of glue, but even then after fixing the glue,
resizing, and unfixing the glue things would "snap" back, I'm not sure
we want this.  And again, all these seems like such a corner case to be
worth the development time.  For now I will just leave this mega commit
message for me to remember these were known issues and I had ideas of
how to solve them.

In the end if these issues become an issue the best solution would be to
actually design the key dependency algorithm to correctly emulate the
gravity to the right.