Fix computation of edge start
I had assumed the sdge start was the previous resized_key, this is
wrong. What we actually want is the last sized segment previous to the
current resized_key.

There is a detail here, that for the case of start being NULL to be
handled, the end key must be updated when resized_key changes.