Simplify key splitting behavior
Change the behavior of switching the position of the created key in the
split when moving accross the middle of the key. Instead, now we don't
change the position but place the new key depending on which edge was
closer hen clicked. Similar to the way we resize the right or lef edge
depending which one is closer.