Fix split_rect_x computation for non rectangular splits
split_rect_x should be set to the x coordinate of the left border of the
segments being splited. For the case of non rectangular splits, this is
different than the x position of the rectangle being clicked. It's the
coordinate of the start edge segment. This commit adds a function to
compute this coordinate.