Change the move key tool for an add key tool
After some thought I think the move key tool is useless. The same
modifications can be done by using the tools: delete key, keycode assign
and add key.

I haven't came up with a good UI for moving keys. Leaving the moved key
in place while dragging makes weird to determine at which point to move
the key. Moving a key may cause a lot of changes that may confuse the
user. There are no defined semantics as to how we should handle keys
that are left floating after picking up a key: move them left? fix them
by creating user glue?.

Instead of drawing the full key being moved it could be removed and a
thin marker could be shown in place, but this is weird for multirow keys
and does not allow to set the user glue when the key is inside a wider
gap.

In the end, I don't think moving keys is very useful, and instead has a
lot of implementation details and design decisions that make it very
complex. Instead I will focus in implementing the add key tool.