Change get_multirow_segment_width () by get_sgmt_width()
The old function get_multirow_segment_widthkey() asserted
key->type == KEY_MULTIROW_SEGMENT but every place it was called we were
using the same idiom of checking the type and if it wasn't a multirow
segment then use key->width. The new version of the function works for
all types of segments and removes all the repeated ifs.