Simplify glue computation while adding a key to the left
Adding a new key beyond the left margin had several quirks and the code
was hard to follow. The new code is simpler and appears to handle all
cases.

We add a new concept of a key being the first: A key is first if 1) Its
multirow parent is the first segment in the row and 2) it's not multirow
or it's multirow and has no other key/segment to the left.

This affects the computation of the glue of the key next to the one
being added (keys added to the right don't need this computation). If a
key is not first, the computed user glue for the key after is clamped
between 0 and the old user glue value. If a key is first, then we only
check the lower bound of 0 and then assign it unconditionally.