Settle on the behavior of key addition beyond the left margin
Previously I thought a complex dependency algorithm was going to be
necessary, to know where to add user glue to move the keyboard right
when adding a key ito the left. Turns out this is not necessary, the
result will always be to move all first keys (see previous commit).

This seems to be the correct solution. It solves the problem of the
incorrect keyboard width computation when allowing negative user glues.