Complete edge cases of segment resize
Solve all edge cases for the segment resize tool and add some
documentation about the reasoning behind the limitation of not having
arbitrary alignment in multirow keys. At the moment I think right and
left alignment are enough for most usecases but maybe in the future we
will want to change that.