Simplify logic of resize segment handling
The implementation of segment resizing beyond the left margin had some
issues, this is a more complete and simpler implementation of it. It
removes some unnecessary glue computations, and fixes some edge cases.

Still there are some issues if the segment being resized changes the
width of the keyboard from the left from its initial position.