BIG simplification of kv_adjust_glue() logic
The conditions inside kv_adjust_glue() were very contrived, after some
careful thought of all cases, I think this is the correct logic for it.

User glue of a key has to change only if the segment where the glue will
change is the only 'supporting segment'. A key segment S is a supporting
segment, if it collides with another segment, when the user glue for the
key containing S is 0. From the definition, a supporting segment of a
multirow key is a segment for which its internal glue is equal to the
key's user glue. Also from the definition, the segment of a non multirow
key is a supporting segment.

With this new condition to change the user glue of a key, we now handle
correctly when the contact point of a multirow key moves from one
segment to another. While having the behavior of non multirow keys as a
specific case of the more general one.