Handle long increments in kv_adjust_glue()
The previous code implemented the simplest idea of the supporting
segment concept. And shows how it is relevant when changing the glue of
a segment from a multirow key. But it assumed increments were always the
smallest possible, and that at some point the glues of two segments
would be exactly the same. In practice this is not the case.

The new function multirow_user_glue_for_delta_glue() handles the case
when the change in the glue is so big, that it was never the case that
two segments of a multirow key had the same total glue.

Also, the previous implementation required some fixing at the boundary
condition. This also fixes that.