Change meaning of internal_glue in non parent multirow segments
Previously the glue computation included a key's user glue in each of
its segment's internal_glue. This meant that computing the "real"
internal glue for a segment, required subtracting the key's user glue
from the internal_glue property of the segment. This was incorrectly
used in some places like key rendering, where we assumed every segment's
total glue was the sum of it's user glue and it's internal glue.

This commit changes the meaning of intenal_glue, in non parent multirow
segments to be the "real" internal glue. Making the total glue, of any
type of key segment equal to internal_glue + user_glue.

This choice has a very small performance cost, because now rendering a
multirow segment iterates over the multirow key to get the parent and
from it the user_glue (something we were supposed to be doing long ago
but weren't really doing). If this ever becomes an issue, we may cache
this while rendering, or make parent lookup be O(1) by adding a pointer
to it in each non parent multirow segment.

This also simplifies the detection for supporting segments. A segment is
a supporting segment if its internal_glue == 0. Removing the distinction
between parent and non parent segments inside a multirow key.