Change behavior of glue adjusting on segment resize
Make resize segment be 'memoryless'. Previously we were adjusting the
glue for the resized segment almost unconditionally (actually I think it
should have been done unconditionally and it was a bug). This had the
effect of always trying to keep the other keys fixed in place.

When the segment got big and consumed all the glue available, inevitably
it pushed keys to the right. Then when shrinking the segment (without
having clicked to accept the change), the pushed keys would be left
in it's pushed position, not allowing the user to go back to the
starting position. I say this type of behavior has 'memory'. I think
this should not happen and the user should go back to the starting
position to cancel the changes made.

This commit now adjusts glue on a very specific condition. 1) The glue
is getting smaller or 2) the glue is increasing and the size of the
segment equal or less than the original width plus the original glue and
the glue is not grater than the original glue.o

This is a mixture of two behaviors. One where the glue always changes
and another one where the glue never changes. Here we change the glue
but don't allow it to grow beyond the value it had before. Time will
tell how useful and intuitive this behavior really is.