Change glue adjustment in segment resize (again)
The previos implementation was computing the total glue in the segment
where it should have been in glue_key. Fixing this showed that the
previous condition wasn't very good. Resizing a segment that had user
glue would keep the next key fixed while the glue got smaller (consuming
the glue), but it moved the next key when the segment got smaller
because we didn't allow the glue to grow in size.

A new idea that I'm exploring now is that users expect the glue to
grow/shrink if there is already glue, but they don't expect glue to
appear if it was not there before.

Still, a case where I find this behavior to be weird, is when resizing non
multirow keys' left edge. I think rectangular keys should adjust the
glue unconditionally and resizing a the segment of a non multirow key
should behave in this way.