Add tools to extend/shrink keys vertically
Up to now there was no way to create multirow keys with tools from the
keyboard view. This commits adds two separate ones, one that extends a
key vertically and another one that shrinks it.

The intended behavior is that they act inmediately after clicking a key.
If the click was made in the upper half of the key then the key grows
from above, otherwise it grows from below.

Another behavior could have been one like the one for horizontal key
resizing (click to start, move mouse, click to end), but it seems much
more complex and produces states in the UI I'm not sure how to handle.

What happens if the top of a key is dragged beyond the top/bottom of the
keyboard?
    a) Do nothing.
    b) Create new row.
    c) Create new row and then modify its width.

What happens if the top/bottom of a non rectangular multirow key is
shrunken?
    a) Do nothing.
    b) Shrink the key and when it's extended back create a
       rectangular key.
    c) Shrink the key and store the widths of the key so it can be
       restored after the key is extended back.

How is the glue affected by extending a key?
    a) Adjust the glue to try to keep the position of the key fixed.
    c) Don't adjust the glue and allow things to move around.

    Note: In any case, the glue state has(?) to be stored so the user can
    get back to the same starting position, and we don't know a priori the
    number of rows which requires dynamic allocation(s).

At the moment, I've chosen to implement the simpler version even though
the tools added are not similar to existing tools for horizontal
resizing. Maybe later I will have good anwers to the questions, and
change the tools to get a more uniform UI.

This commit implements shrinking at the bottom of the keys and has a
stub for extending at the bottom of the key.