Adjust glue while shrinking a multirow key
We now adjust the glue for key shring similarly as segment resizing. If
there is no glue, then we don't add glue and allow thing to move, if
there is glue then we adjust it to keep everything fixed.

There is a special case when a multirow key becomes non multirow because
these keys must have internal glue of 0. In this case we move its
internal glue into the user glue of the remaining segment.