Handle vertical extend when the vertically adjacent segment is multirow
When extending a key, either from the top or from the bottom, if the row
already exists, we compute the fist segment S such that its right edge
is to the right of the left edge of the edge segment. Normally we put
the new segment before S, except in the case where S is multirow and
another segment of S's key is located to the left of the key being
extended. In this case, we must add the new segment after S and not
before.

This commit also contains simple style changes to make it more concise.