Correctly adjust the left edge when resizing the left edge of a segment
Correct handling of moving a left edge beyond the left margin is more
complex than previously thought. Missing cases included when the change
in width went through the left margin, or when reching the left margin
exactly from the left.

New cases are based on comparing the order of the values for old segment
width (a), new segment width (b) and the distance of the left margin
from the right edge of the segment (c). This creates 6 possible
permutations, handled in 4 different ways:

 - Do nothing. c>a>b || c>b>a
 - Adjust left edge by a-b. a>b>c || b>a>c
 - Adjust the left edge by c-b. a>c>b
 - Adjust left edge by a-c. b>c>a

In the code the a, b and c values are represented by:

    c = original_glue_plus_w
    b = sgmt->width
    a = sgmt->width - delta_w

Which implies delta_w = b-a.

Edge cases casued when either a, b or c have equality are also handled
by this code.