Fix wrong glue adjustment on big segment width changes
Segment width changes sometimes adjust glue, sometimes not, depending on
if the edge being resized is after or before the previous user glue.
This is done to always leave the key where it started, no matter how far
to a side it may have been pushed while resizing.

This causes a problem if a very big resize is done. If the resized edge
had pushed a multirow key to one side moving it by an amount we call
"dead glue", then a very big step that shrinks the segment back so much
that it stops being the supporting segment, will compute the glue as if
it was still a supporting segment. This happens because the next minimum
glue of the multirow key is computed including the dead glue, and it may
be the case that for this glue the sgment is still "supporting". In the
end the computed user glue will be wrong by the amount it moved after it
stopped being the supporting segment.

To solve this issue the width change is split into two smaller ones, one
that removes the dead glue, and a second one that actually computes the
correct glue. I'm not 100% happy with the solution but it works for now.