Definitive fix for big steps in segment resize
The previous commit had the correct solution, but was incorrectly
implmented. This commit abstracts the part of kv_change_sgmt_width()
that should be called in the two stages of handling this specific case
into the function kv_resize_sgmt().

Also documentation was added to explain the edge cases that make the
code look the way it looks.