Fix multirow case in kv_adjust_left_edge()
The expected behavior of this function is to create the illusion of
things moving to the left of the left margin by moving everything else
to the right. The row where things are going beyond the left margin is
the one containing the provided sgmt argument as its first segment.

If sgmt was par of a multirow key, but not the multirow parent, things
were not working correctly because we were only looking at key parents
of first keys. What this commit does is check for all segments in the
multirow key and check if it contains sgmt, if it does, then that
key parent is not modified.