Update the keycode lookup table where necessary
The array kv->keys_by_kc stores pointers to the parents of keys indexed
by keycodes. This means that whenever the key parent changes, an update
must be made to this table. This update was missing while removing a
key, and when resizing a key from the top, either shrinking or extending.