Complete the new implementation of the edge resize tool
The case of non rectangular edge resizing was: whatever happened after
resizing the key. The reason being, that I didn't had a clear idea of
how this should work. This is an attempt at implementing a behavior
which I think is consistent with the other tools.

This commit and the previous one, implement the kv_change_edge_width()
function. The intended behavior of this function is to allow an edge to
push other keys to the side, in such a way that going back to the
original width will leave them scattered at their original positions.

Changed the test geometry to one that better shows the behavior of this
function.