Simplify logic of kv_locate_edge()
A lot of the logic in kv_locate_edge() seemed to be cruft from the time
where I was still unsure about how to represent an edge. Now that I have
a better understanding of how an edge should be representet, I didn't
understand most of the logic here.

Also, the implementation was buggy. In the new test geometry, resizing
the second right edge, by clicking a different segment than the second
of the key caused kv_locate_edge() to return incorrect values and after
some resizing a crash.

Finally, the code to debug this function had bit rotted and caused
compilation problems. I updated that too.