small improvements to kv_change_edge_width() and save_edge_glue()
We were having a crash (stack overflow) when resizing a right edge that
didn't have any keys to the right. Although the changes done to common.h
in the previous commit fixed that, these commit adds other related
changes that make code more robust.

Don't allocate the glue info array in save_edge_glue() before knowing
the actual number of items it will have. We allocate a temporary array
in the stack and then if something was stored in it we save it to one
allocated inside the pool which has the correct size.

In kv_change_edge_width() check we have a glue_info, if we don't then
don't do anything strange.