Make save_edge_glue() handle segments with no keys to the side
When a right edge does not have a key to the right then we still need to
add a glue info, otherwise all num_sgmts don't add up to the length of
the edge. We set min_glue=0 to these items and let kv_adjust_glue()
figure out there is no key, and that nothing should be done.

If all created items have min_glue==0 then either there is nothing to
the right of the edge, or there is a key that is fully attached to the
edge, in this case we don't do anythyng as the intuitive thing felt to
not make glue appear magically from thin air.