Better implementation of kv_adjust_left_edge()
I now think the "first key" concept is flawed. A key may be touching the
left edge and thus require a user glue update when adjusting the left
edge but at the same time not be a first key.

This new approach lists all parents of keys that are "touching" the left
edge (are supporting segments and first segment of a row). From these we
discard those of which the left margin supporting segments are not the
only ones or the one explicitly discarded by the sgmt argument. Finally,
all remaining keys are pushed to the right (user glue is incremented),
while avoiding pushing them multiple times even if they contain multiple
supporting segments touching the left margin.