Remove anchors from resize functions (x_anchor, y_anchor)
There were two different (but similar) patterns to handle the
computation of a resize. On one hand we had absolute computations that
used the clicked position, computed the delta to the current pointer
position and set the new size to the original size plus this delta. The
second approach was a differential one, where we had an extra point
called an anchor, this point was updated to be the position of the last
update, then a delta was computed from here and the resize happened by
this value. This second approach adds more complexity as the anchor
update depends on the side the update is happening on.

Even though the second approach was correct, it was used at a time where
I thought some sizes weren't well defined. I thought width of the edge
of a non rectangular key was "meaningless", in practice we defined the
width of an edge as the width of the start segment. In a similar way for
segment resizing, I thought it was meaningless for non sized segments,
but we now store the width in the key_t structure for all cases.  This
makes the width of all resizeable things well defined, which in turn
makes differential updates (and anchors) unnecessarily complicated.

This commit removes anchors, makes all resizes absolute, and makes code
as similar as possible where the same pattern is used.