Start implementation of a new algorithm to adjust glue for an edge
I has became apparent that the glue adjustment should happen either per
segment or per edge. Running the segment version multiple times on an
edge update is not enough, a new algorithm similar to the one used right
now for left edge adjustments seems to be the right approach.

I think when finished, this algorithm should actually replace the one
for left edge adjust.

The main issue had to do with what constitutes next_min_glue. I was not
able to define it in such a way that would both work for segment glue
adjustments, edge adjustments. Here I went back to an old version that
worked fine for segment adjustments, but failed on edge adjustments. The
details on how and why these failed are very subtle, so I won't go into
detail but by the end of this reimplementation I hope to have several
test cases to know the algorithm handles all of them.