Fix logic for left edge adjustments in edge and segment resize
The old way in which we computed if an adjustment of the left edge was
required was wrong for both cases because we didn't take into account
the case when the edge collided with the left edge. Common code was
abstracted and used both for segment and edge resize tools.

Also abstracted out the function to compute an edge's length.