kv_change_edge_width() and kv_change_sgmt_width() cleanup
Bring back the functionallity that was commented in
kv_change_sgmt_width(). Now that we got to a working version of
kv_change_edge_width(), we can make both functions mirror each other so
that it's easy to compare the similarities and differences between them.

Remove debug info that was being used to develop kv_change_edge_width().

Delete redundant (and incorrect) edge adjustments that happened when
cancelling a segment resize. That is now correctly handled in
kv_change_sgmt_width().