Attempt at cleaning up the tooltip code
In summary even though I solved the previous issue, I now ran into a
2 year old Gtk bug about tooltips on buttons inside an overlay only
showing when hovering the border, because we use a flat style the border
is 0px which means we get no tooltips (and showing tooltips when
hovering the border would be equelly useless).

I tried to work around this by overlaying a GtkDraw area on top of the
toolbar, but this then eats all button presses/releases and makes the
toolbar useless. I'm out of ideas but I don't want to throw away the
code in case sometime in the future the Gtk bug gets fixed, I added a
define to switch between my workaround and the clean code to easily test
if this has been solved.

Lets hope the code doesn't rot so badly...