common.h: Improve path manipulation functions
Make them use pom macros to let the user call them with NULL as pool and
have the result allocated with malloc.

Added the function remove_multiple_extensions() to easily remove
multiple extensions like ".tar.gz". It is careful about allocating just
the result.

Add path_split() to avoid using basename() and dirname() which have
multiple versions that differ on how things are allocated and if they
change the input. path_split() never changes the input and the result
must be freed either by destroying the pool that was passed, or by
calling free() if NULL was passed as pool.