Add keyboard view representations
A keyboard view representation is used to store the string corresponding
to how keys are located in a layout, and which keycode is assigned to
each one of them. Because the description of a keyboard's geometry can
be done either with a function or by parsing a file, this new code will
unify all of them into the same format.

We will now be able to distinguish and handle differently predefined
internal geometries, user geometries, files created for autosave, debug
geometries created with functions, etc.

Also, I moved the code that converts a keyboard view from and to a
string into it's own file.