Refactor how we handle current representation string
Previously the current string representation of the keyboard was stored
in a string that was malloc'd and freed individually. This was redundant
information as the string found inside the representation should have
this. We were letting both of these go out of sync and syncing them when
saving the representation, which regenerated the representation store.

Instead we now keep track of all string representations of every change.
This is kept inside the representation as a linked list. This new
architecture makes pointless the is_saved flag, as changes made after
a load are not stored as a different representation but as a new state
inside this representation. It also avoids having representation_str
keeping the current string representation as the latest state of the
current representation. Finally, this lays the groundwork for a future
implementation of CTRL+Z.