Clamp added_key_user_glue to 0 if locate_stat == LOCATE_HIT_KEY
Unfortunately we were allowing added_key_user_glue to be negative when
the internal glue was less than 1. Then we detected this as a key being
added to the left of the keyboard and were calling kv_adjust_left_edge()
with this value, effectively shifting the layout when it was unnecessary.

Also we were only taking into account the internal glue when we actually
need the total glue. This bug came from a time where user glue didn't
exist.