Also load internal representataions from gresource
Sometimes it's easier to create a representation using our tools than
writing a function for it, for instance in the case of the Full layout
(laptop layout with a keypad). This commit allows to also load internal
representations from .lrep files inside gresource, and adds the
Full.lrep file in such way.