Set keycode to 0 when a key becomes unassigned
The string created by kv_to_string() and parsed by kv_set_from_string()
represents unassigned keys, as keys with keycode of 0, they don't care
about the KEY_UNASSIGNED type. This caused an issue when reassigning a
keycode from a saved representation.

Because we create an autosave and then reload representations, setting
the current representation to the one comming from the new autosave
file, the key that was previously unassigned but left with its old
keycode, would become of type KEY_DEFAULT, and have the same keycode as
the new one, creating a layout with 2 keys that show the same keycode,
which should not happen.

This issue also showed up when reassigning keycodes, leaving some keys
unassigning, closing the app and openning it again. All unassigned keys
would become assigned and keep the old keycode, also creating a layout
with multiple keys that have the same keycode.

The type KEY_UNASSIGNED is redundant if we define a keycode of 0 to mean
unassigned, maybe we should remove it in the future.