Specialize kv_to_string() into 2 versions
For debugging purposes we usually want to dump as much information as
possible, specially information that has been internally computed, this
commit now creates a new kv_to_string_debug() function for this usecase.
On the other hand for the more common case of storing the state as a
string, we don't want redundant or unnecessary information,
kv_to_string() now doesn't print any of this information.