Don't load the representation in kv_set_current_repr()
Took the functionality that rebuilds the data structure out from
kv_set_current_repr() and into kv_load_current_repr(). The reason for
this is that reloading the datastructure is not something we want to do
in places we are not expecting it, as this invalidates any pointer into
the data structure.

Not being careful about this caused a bug in the implementation of
multiple keycode assignments. kv_autosave() triggered a data structure
reload when doing the first change to a saved layout. In the case of
multiple keycode assignment we store a pointer to the currently selected
key, call kv_autosave() and expect the pointer to still be available.