Use + and - icons in the list to install/uninstall layouts
I think the - icon in the headerbar wasn't clear enough about what it
was going to remove, also the Install layout button as an introduction
button was not clear about what it did. No these buttons are below the
custom layout list and are + and - indicating thet they add and remove
from the list. The tootltip explains it will install a layout to the
system, hopefuly making clearer what layouts in that list mean.

We want the user to know this list has installed custom layouts, which
is different from active ones. The list of active layouts is the one
found in switchboard. ATM I don't want to duplicate fuctionality from
the keyboard settings panel.