Implement the functionality of the back button
The function set_custom_layout_list() came from a tibe before there was
replace_wrapped_widget(), it made us accidentally use the replace
pattern when building the widget from scratch. I now moved to the
wrapped widget idiom and only use it when updating just the list (a
layout is installed or removed), otherwise the widges is just created
wrapped and set in the the app's main state structure.