Some refactoring of the headerbar UI and welcome sidebar
Abstracted the welcome sidebar so when going back to it we just update
it and the headerbar, leaving the keyboardview untouched, we were
leaking keyboard views before because we were creating new ones without
destroying the old ones. Maybe tie keyboard view destruction to the
destruction of the widget?.

Also refactored the headerbar buttons to use the replace widget idiom
and revisited the grab/ungrab code. At some point we were trying to
detect when the grab was broken externaly and we tried to react to this,
unfortunately this never worked so this code seems useless now.  What we
do now is grab everything making it as hard as possible to break the
grab. Maybe something weird would happen if for example a popup shows
while having a grab, I should chek this in the future.