Set the view libxkcommon state when openning an xkb file.
The keyboard view uses libxkbcommon to compute the state of each key
depending on what the user presses or clicks. Previously we assumed the
xkb file came from the reconstruction of the current layout. Now we pass
it as parameter so we can set it to the content of an xkb file when
opening one, and to the reconstructed keymap when editing an installed
layout.