Fix some bugs when opening and closing keymaps for editing
I was destroying the keymap twice. I now just destroy it before creating
a new one. Previously I was destroying it but from another callback,
it's not obvious that this callback happens always before creating a new
one.

I also now connect the row selection handler before selecting the first
row in the custom layout list. This fixes an issue, but now we call this
handler twice at startup.