Don't complain about multiple modifier definitions
Even though the C syntax of XKB files implies modifier definitions
should be scoped, this seems pretty much useless and only adds
complications when resolving modifiers in the symbols section. Lets not
care about this for now and later see if it actually matters
semantically. If it does matter, we would have to have 2 modifier
definition trees for the types and compatibility sections, and then
merge them somehow.